+++ Program.cs
@@ -167,6 +167,7 @@ namespace DNWS
             } while (ns.DataAvailable);
 
             request = new HTTPRequest(requestStr);
+            Console.WriteLine(requestStr);
             request.addProperty("RemoteEndPoint", _client.RemoteEndPoint.ToString());
 
             // We can handle only GET now
@@ -233,11 +234,6 @@ namespace DNWS
         private static DotNetWebServer _instance = null;
         protected int id;
 
-        public class var1
-        {
-            public static string name = "123";
-
-        }
         private DotNetWebServer(int port, Program parent)
         {
             _port = port;
@@ -263,12 +259,6 @@ namespace DNWS
         /// <summary>
         /// Server starting point
         /// </summary>
-        /// 
-        //public class Variables
-        //{
-        //    public static string name = "123";
-
-        //}
         public void Start()
         {
             while (true) {
@@ -295,12 +285,6 @@ namespace DNWS
                 {
                     // Wait for client
                     clientSocket = serverSocket.Accept();
-                    //
-                    //
-                    
-                    var1.name = clientSocket.RemoteEndPoint.ToString();
-
-                    //
                     // Get one, show some info
                     _parent.Log("Client accepted:" + clientSocket.RemoteEndPoint.ToString());
                     HTTPProcessor hp = new HTTPProcessor(clientSocket, _parent);
@@ -319,3 +303,4 @@ namespace DNWS
         }
     }
 }
+
