+++ config.json
@@ -1,25 +1,32 @@
 ﻿{
     "DocumentRoot": ".",
     "Port": "8080",
-    "Plugins" : [
-        {
-            "Path" : "stat",
-            "Class" : "DNWS.StatPlugin",
-            "Preprocessing" : "true",
-            "Postprocessing" : "false" ,
-            "Singleton" : "false"
-        },
-        {
-            "Path" : "ox",
-            "Class" : "DNWS.OXPlugin",
-            "Preprocessing" : "false",
-            "Postprocessing" :"false" 
-        },
-        {
-            "Path" : "delay",
-            "Class" : "DNWS.DelayPlugin",
-            "Preprocessing" : "false",
-            "Postprocessing" :"false" 
-        }
-    ]
+  "Plugins": [
+    {
+      "Path": "stat",
+      "Class": "DNWS.StatPlugin",
+      "Preprocessing": "true",
+      "Postprocessing": "false",
+      "Singleton": "false"
+    },
+    {
+      "Path": "ox",
+      "Class": "DNWS.OXPlugin",
+      "Preprocessing": "false",
+      "Postprocessing": "false"
+    },
+    {
+      "Path": "delay",
+      "Class": "DNWS.DelayPlugin",
+      "Preprocessing": "false",
+      "Postprocessing": "false"
+    },
+    {
+      "Path": "clientinfo",
+      "Class": "DNWS.ClientInfo",
+      "Preprocessing": "true",
+      "Postprocessing": "false",
+      "Singleton": "false"
+    }
+  ]
 }
