--- newhw3
@@ -1,1 +1,0 @@
-
